(MAINT) Mock PDK::Util::Env.fetch resp for util_spec.rb:292
Prior to this commit, when running this test in the context of a
Github Actions runner, due to the `XDG_CONFIG_HOME` being set,
the PDK::Util::Env.fetch method was returning the value of this
ENV var, given it was set and did not need to return the fallback
value (2nd arg).

This commit now mocks the return PDK::Util::Env.fetch so it is
consistent whether `XDG_CONFIG_HOME` is set or not.