(GH-1083) Bump childprocess to '~> 4.0.0'; Disable @process.leader
Prior to this commit, it was observed on Windows Server 2016, 2019
Github Actions instances that invoking the pdk would result in
ACCESS_DENIED errors.

When the ChildProcess.leader parameter is set to true, it will set
the `CREATE_BREAKAWAY_FROM_JOB` and `JOB_OBJECT_LIMIT_BREAKAWAY_OK`
flags in the Win32. This is not required in Windows > Server 2008 / 7
and can potentially cause issues when Task Scheduler attempts to
schedule a job that invokes a program with insufficient privileges.

This commit also bumps `childprocess` to `~> 4.0.0` and `hitimes`
to `2.0.0` to resolve dependency issues after bumping `childprocess`.