(GH-1113) Fix incorrect opt type for `forge-token`
Prior to this fix, the `forge-token` option defined in the PDK::CLI's
`release` subcommand did not have the `argument:` attribute set.
This meant that the underlying Cri lib treated it as a boolean option
and the forge token value was set to `true` and passed to the
`PDK::Module::Release` object.

This fix sets the `argument:` attribute with the value `:optional`,
which allows it to be optional but treated as an arg that needs the
value extracted after `=`