(GH-917) Fix incorrect 'Missing forge-token' error throwing
Prior to this commit, when running `pdk release` and stepping
through the interview questions, even if you selected `N` to the
publish forge question, a warning would still display stating:

```sh
pdk (ERROR): Missing forge-token option
```

This commit sets the `:skip_publish` option to true in `opts` that
gets passed to the `PDK::Module::Release` object on init. This will
ensure that the `skip_publish?` call returns `true` and the
Forge token check is not performed.

This was incorrectly asserted as resolved in `2.1.1` of the PDK.