(PDK-1758) supplement `in_module_root` with check for `metadata.json`
Prior to this change the `in_module_root` function looked only for specific folders within the project to determine if it was at the root. 
This check supplements it with a check to see if `metadata.json` is present enabling  modules that only contain `/files` without looking for such a common folder name.