(MAINT) Refactor mocks in spec for ruby_version
Prior to this commit the spec for #available_puppet_versions relied
on mocking calls to the file system to get Gem::Specification to
load the stubbed specs; this was brittle and eventually broke when
the upstream library moved to an internal method that called File.open()
instead of File.read().

This commit reworks the spec to instead mock Gem::Specification.load()
directly, reducing fragility and keeping us from mocking the internals
of an upstream library.