(CONT-487) - Fix bundler test
In this commit we have reworked the interactive test to use PTY.spawn.

Additionally this test has been scoped to only execute when the following
conditions are true:

  * We are not on a windows platform
  * The current active ruby version is lower than 2.7

As of today, it does not seem possible (within reason) to properly
send data to the processes write stream when running on ruby 2.7.

This causes an issue where the tests will hang indefinetly when we
try to retrieve output (r.gets) from the process because there is nothing
there to retrieve.

After a short discussion internally we decided to scope this test
especially as this feature may change in the PDK in coming releases.