(CONT-755) Fixup schema validation
Prior to this change schema validation was broken with the latest
version of the json-schema gem. This is because the v6 schema name has
change from draft-6 to draft.

This change updates references to the v6 schema. Additionally we now
include the task schema so that we can update the schema name.