(CONT-1026) Deprecate release command set
The release prep command introduces a dependency on the github-changelog-generator
gem which is unmaintained and broken.

This ticket exists primarily to remove this command but that introduces
an unnecessary level of commands for users. For example

```
pdk release
pdk release prep
pdk release publish
```

This change deprecates the `release` command set. They will be removed
in a future version.