(CONT-1027) Mark --pe-version as deprecated
The --pe-version flag has not been fully functional since PDK 2.7.
This is because the metadata it used to derive versions became out
dated and caused inconsistencies.

Additionally we also remove a depependency on a third party service.

This change marks the option deprecated and adds a warning to
direct users towards the --puppet-version flag