(CONT-1154) Disable console command
The console command relies on puppet-debugger which currently does not
support Ruby 3 and Puppet 8.

Given that the command has always been an experimental feature of PDK,
a decision was made to disable it for this release.
This will give us time to work on some updates for a later release.