(CAT-648) - Mark release*.rb files as non-executable
Prior to this commit, the release.rb and release_spec.rb files were
marked as executable. This was an error, and these files do not need to
have these permissions. This commit updates the permissions of these
files to match the other files found in the ./lib/pdk/cli and
./spec/unit/pdk/cli directories.