(CAT-2122) Temporarily exclude the windows 2019, ruby 3.2 acceptance test.
Our current daily nightly summary includes a failure on the pdk acceptance test that has been occurring for some months, specifically the windows 2019, ruby 3.2 one.  Although we are aware that this error is indeed outstanding and occurs every day, the problem is that it can easily “hide” a valid error that we need to fix.  This PR temporarily disables this test until the CAT-2122 ticket addresses and fixes the test.   When fixed, then this PR will be rolled back.

Signed-off-by: Gavin Didrichsen <gavin.didrichsen@gmail.com>